+++ dst/routes/transactions.js
@@ -84,19 +84,10 @@ transactionsRouter.all('/start', (req, res, next) => __awaiter(void 0, void 0, v
             project: { id: { $eq: req.project.id } },
             inCodeSet: { identifier: { $eq: chevreapi.factory.categoryCode.CategorySetIdentifier.AccountType } }
         });
-        const productService = new chevreapi.service.Product({
-            endpoint: process.env.CHEVRE_API_ENDPOINT,
-            auth: req.user.authClient
-        });
-        const searchPaymentCardsResult = yield productService.search({
-            project: { id: { $eq: req.project.id } },
-            typeOf: { $eq: 'PaymentCard' }
-        });
         res.render('transactions/start', {
             values: values,
             message: message,
-            accountTypes: searchAccountTypesResult.data,
-            paymentCards: searchPaymentCardsResult.data
+            accountTypes: searchAccountTypesResult.data
         });
     }
     catch (error) {